fix: fixed aino's stuff <3
