+++ .eslintrc.json
@@ -3,7 +3,9 @@
 	"rules": {
 		"@typescript-eslint/no-throw-literal": 0,
 		"@typescript-eslint/restrict-plus-operands": 0,
-		"@typescript-eslint/naming-convention": 0
+		"@typescript-eslint/naming-convention": 0,
+		"@typescript-eslint/no-floating-promises": 0,
+		"@typescript-eslint/no-namespace": 0
 	},
 	"overrides": [
 		{
