+++ package.json
@@ -2,7 +2,7 @@
   "name": "sapphire",
   "version": "0.0.0",
   "description": "testing sapphire and discord",
-  "main": "src/Sapphire.ts",
+  "main": "dist/Sapphire.js",
   "author": "KunoichiZ",
   "license": "Apache-2.0",
   "private": true,
@@ -22,34 +22,48 @@
     "watch": "tsc-watch -b src --onSuccess \"node dist/Sapphire.js\""
   },
   "devDependencies": {
-    "@types/node": "^14.14.14",
+    "@favware/graphql-pokemon": "^5.0.0",
+    "@sapphire/eslint-config": "^1.1.0",
+    "@sapphire/prettier-config": "^1.0.3",
+    "@sapphire/ts-config": "^1.0.3",
+    "@types/node": "^14.14.16",
     "@types/node-fetch": "^2.5.7",
+    "@types/ws": "^7.4.0",
+    "@typescript-eslint/eslint-plugin": "^4.11.0",
+    "@typescript-eslint/parser": "^4.11.0",
+    "eslint": "^7.16.0",
+    "eslint-config-prettier": "^7.1.0",
+    "eslint-plugin-prettier": "^3.3.0",
+    "prettier": "^2.2.1",
     "tsc-watch": "^4.2.9",
     "typescript": "^4.1.3"
   },
   "dependencies": {
-    "@favware/graphql-pokemon": "^5.0.0",
-    "@sapphire/decorators": "^1.0.2",
-    "@sapphire/eslint-config": "^1.0.4",
-    "@sapphire/event-iterator": "^1.1.0",
-    "@sapphire/framework": "^1.0.0-alpha.0",
+    "@sapphire/decorators": "^1.1.0",
+    "@sapphire/event-iterator": "^1.1.1",
+    "@sapphire/framework": "github:sapphire-project/framework#build",
     "@sapphire/pieces": "github:sapphire-project/pieces#build",
     "@sapphire/stopwatch": "^1.0.1",
-    "@sapphire/time-utilities": "^1.2.0",
+    "@sapphire/time-utilities": "^1.2.1",
     "@sapphire/type": "^1.1.1",
-    "@sapphire/utilities": "^1.1.1",
+    "@sapphire/utilities": "^1.2.0",
+    "bufferutil": "^4.0.2",
     "chokidar": "^3.4.3",
     "colorette": "^1.2.1",
     "common-tags": "^1.8.0",
     "discord.js": "^12.5.1",
-    "eslint": "^7.16.0",
     "node-fetch": "^2.6.1",
-    "typeorm": "^0.2.29"
+    "pg": "^8.5.1",
+    "typeorm": "^0.2.29",
+    "utf-8-validate": "^5.0.3",
+    "ws": "^7.4.1",
+    "zlib-sync": "^0.1.7"
   },
   "imports": {
     "#lib/*": "./dist/lib/*.js",
     "#orm/*": "./dist/lib/orm/*.js",
     "#root/*": "./dist/*.js",
     "#utils/*": "./dist/lib/util/*.js"
-  }
+  },
+  "prettier": "@sapphire/prettier-config"
 }
