+++ src/Sapphire.ts
@@ -1,54 +1,25 @@
 import 'reflect-metadata';
-import { TOKENS, PREFIX, PGSQL_ENABLED } from '#root/config';
-import { SapphireClient } from '@sapphire/framework';
-import { Guild, Message } from 'discord.js';
-import type { ClientOptions } from 'discord.js';
-import GuildSettings from '#lib/orm/entities/GuildSettings';
-import GuildSettingRepository from '#lib/orm/repositories/GuildSettingRepository';
-import { getCustomRepository } from 'typeorm';
+import { connect } from '#lib/orm/ormConfig';
+import { SBClient } from '#lib/SapphireClient';
+import { PREFIX, TOKENS } from '#root/config';
 
-export class SBClient extends SapphireClient {
+const main = async () => {
+	const client = new SBClient({
+		defaultPrefix: PREFIX,
+		presence: {
+			activity: {
+				name: 'Pokémon Emerald',
+				type: 'STREAMING'
+			}
+		}
+	});
 
-    public constructor(options?: ClientOptions) {
-      super(options);
-      this.registerUserDirectories();
-      this.fetchPrefix = async (message: Message) => {
-        if (!message.guild) return [PREFIX, ''];
-        return this.fetchGuildPrefix(message.guild);
-      };
-  }
-
-  public settingsCache = new Map<string, GuildSettings>();
-  
-  public get invite() {
-		return `https://discord.com/oauth2/authorize?client_id=${this.user!.id}&scope=bot`;
-  }
-
-  public async fetchGuildPrefix(guild: Guild) {
-		if (!PGSQL_ENABLED) return PREFIX;
-		const guildSettings: GuildSettings = await getCustomRepository(GuildSettingRepository).ensure(this, guild);
-		return guildSettings.prefix;
+	try {
+		await connect();
+		await client.login(TOKENS.BOT_TOKEN);
+	} catch (error) {
+		client.logger.error(error);
 	}
-  
-  private _version = [1, 0, 0];
+};
 
-  public get version() {
-		const versionStr = this._version.join('.');
-    return versionStr;
-  }
-}
-
-const client = new SBClient({
-    defaultPrefix: PREFIX,
-    presence: {
-      activity: {
-        name: 'Pokémon Emerald',
-        type: 'STREAMING'
-      }
-    }
-  });
-  
-client.login(TOKENS.BOT_TOKEN)
-    .catch((error) => {
-      client.logger.error(error);
-});
+main();
