+++ src/commands/Misc/ping.ts
@@ -8,11 +8,9 @@ import SapphireCommand from '#lib/SapphireCommand';
 	description: 'Runs a connection test to Discord.'
 })
 export default class PingCommand extends SapphireCommand {
-
 	public async run(message: Message) {
 		const sent = await message.channel.send('Pinging...');
 		const ping = sent.createdTimestamp - message.createdTimestamp;
 		return sent.edit(`Pong! That took ${ping}ms. Latency: ${this.context.client.ws.ping}ms`);
 	}
-
 }
