+++ src/tsconfig.json
@@ -1,34 +1,15 @@
 {
-    "compilerOptions": {
-		"target": "ES2020",
+	"extends": "../tsconfig.base.json",
+	"compilerOptions": {
 		"outDir": "../dist",
 		"rootDir": ".",
 		"baseUrl": ".",
-		"module": "commonjs",
-		"moduleResolution": "node",
 		"paths": {
 			"#utils/*": ["lib/util/*"],
 			"#lib/*": ["lib/*"],
 			"#root/*": ["*"],
 			"#orm/*": ["lib/orm/*"]
 		},
-		"composite": true,
-		"declaration": true,
-		"removeComments": true,
-		"sourceMap": true,
-		"inlineSources": true,
-		"esModuleInterop": true,
-		"emitDecoratorMetadata": true,
-		"experimentalDecorators": true,
-		"allowJs": true
-	},
-	"watchOptions": {
-		// Use native file system events for files and directories
-		"watchFile": "useFsEvents",
-		"watchDirectory": "useFsEvents",
-	
-		// Poll files for updates more frequently
-		// when they're updated a lot.
-		"fallbackPolling": "dynamicPriority"
-	  }
+		"composite": true
+	}
 }
