+++ tsconfig.eslint.json
@@ -1,0 +1,4 @@
+{
+    "extends": "./tsconfig.base.json",
+    "include": ["src", "ormconfig.js"]
+}
